Add margin to threshold. Make time per target textbox in TestTab accept floats.

+++ src/gui/main_window/test_tab/TestTab.py
@@ -16,7 +16,7 @@ class TestTab(Frame.Frame):
             OptionMenu.TargetChoosingMenu(self, c.TEST_TAB_STANDBY, 3, 1, (c.TEST_TARGET_NONE,), columnspan=2),
             Textboxes.LabelTextbox(self, c.TEST_TAB_TOTAL_TIME, 4, 0, command=int, default_value=1, default_disability=True, default_disablers=[c.TEST_TAB_UNLIMITED]),
             Checkbutton.Checkbutton(self, c.TEST_TAB_UNLIMITED, 4, 2, columnspan=2, command=self.enableTotalTime, default_value=1),
-            Textboxes.LabelTextbox(self, c.TEST_TAB_TIME_PER_TARGET, 5, 0, command=int, default_value=9, default_disability=True, default_disablers=[c.TEST_TAB_TARGET_OPTION_MENU]),
+            Textboxes.LabelTextbox(self, c.TEST_TAB_TIME_PER_TARGET, 5, 0, command=float, default_value=9, default_disability=True, default_disablers=[c.TEST_TAB_TARGET_OPTION_MENU]),
             Textboxes.LabelTextbox(self, c.TEST_TAB_TIME_PER_TARGET_PLUS_MINUS, 5, 2, command=int, allow_zero=True, default_value=2, default_disability=True, default_disablers=[c.TEST_TAB_TARGET_OPTION_MENU]),
             Textboxes.ColorTextboxFrame(self, c.TEST_TAB_COLOR, 6, 0, default_value="#ffffb3"),
         ))
